Moved ResourceFileInfo into Models namespace
